Restructure to not use composeWith.
I had a number of issues getting external sub-generators to work the way
I wanted them to work.  Prompts and file management between them was too
difficult or eliminated a lot of the sub-generator value.  So, instead,
I have decided to just implement as much as possible within this
generator.